refactor: move `dagger install` to workspace.go
Move moduleInstallCmd and its flag registration from module.go to
workspace.go, grouping all workspace management commands in one file.
Remove unused analytics import from module.go. Fix playground script
line continuation after rebase.
